fix(usuario): histórico devolve as revisões (tipo de revisão do Envers é int)
O Envers numera revisão como int (DefaultRevisionEntity / REVINFO.rev), mas o
RevisionRepository e o RevisaoUsuarioDTO estavam tipados como Long — a leitura
do histórico estourava ClassCastException (Integer -> Long) e a modal não
carregava. Passa a Integer. Teste de integração cobre o caminho real com Envers
(a fatia de controller mockava o service).

+++ src/main/java/br/com/diegocordeiro/dscproject/dto/usuario/RevisaoUsuarioDTO.java
@@ -15,7 +15,7 @@ import java.time.LocalDate;
 @Getter
 public class RevisaoUsuarioDTO {
 
-    private final Long revisao;
+    private final Integer revisao;
     private final String tipo;
     private final Instant data;
     private final String autor;
@@ -27,7 +27,7 @@ public class RevisaoUsuarioDTO {
     private final String generoDescricao;
     private final LocalDate nascimento;
 
-    public RevisaoUsuarioDTO(Revision<Long, Usuario> revision) {
+    public RevisaoUsuarioDTO(Revision<Integer, Usuario> revision) {
         this.revisao = revision.getRequiredRevisionNumber();
         this.tipo = traduzirTipo(revision.getMetadata().getRevisionType());
         this.data = revision.getMetadata().getRequiredRevisionInstant();
